Tests/LibWeb/Text/expected/WebAnimations/misc/animate-with-mixed-percentages.txt

box is moving in the correct direction: true
